projeto_wayne/requirements.txt
flask
flask-cors
pyjwt